First version of fighting. Not working completely because of stupid other packets being sent(move packets) Also keeps doing lots of attackEnd even though it shouldn't

+++ Server/GameServer/MobActivity.h
@@ -88,7 +88,7 @@ typedef struct _CreatureData
 	float			Walk_Speed_origin;
 	TBLIDX			Name;
 
-	WORD			CurLP;
+	DWORD			CurLP;
 	WORD			CurEP;
 	WORD			MaxLP;
 	WORD			MaxEP;
