+++ Server/GameServer/MonsterClass.h
@@ -72,7 +72,7 @@ public:
 		float			Walk_Speed_origin;
 		TBLIDX			Name;
 
-		WORD			CurLP;
+		DWORD			CurLP;
 		WORD			CurEP;
 		WORD			MaxLP;
 		WORD			MaxEP;
